done with rest apis

+++ controllers/feed.js
@@ -2,9 +2,10 @@ const { validationResult } = require("express-validator");
 const Post = require("../models/post");
 const fs = require("fs");
 const path = require("path");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
-  let totalItem;
+  let totalItems;
   const currentPage = req.query.page || 1;
   const perPage = 2;
   Post.find()
@@ -44,7 +45,6 @@ exports.createPost = (req, res, next) => {
   }
 
   const imageUrl = req.file.path;
-  console.log(imageUrl, 1111);
   const title = req.body.title;
   const content = req.body.content;
   // Create post in db
@@ -52,16 +52,25 @@ exports.createPost = (req, res, next) => {
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: "bhojraj",
-    },
+    creator: req.userId,
   });
+  let postData, creator;
   post
     .save()
     .then((result) => {
+      postData = result;
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post created successfully!",
-        post: result,
+        post: postData,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -118,6 +127,12 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -147,8 +162,20 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
+    })
+    .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
     })
     .then((result) => {
       res.status(200).json({ message: "Post deleted" });
